Refactor GUI into module
