+++ src/main.rs
@@ -32,11 +32,12 @@ use rayon::prelude::*; // Add Rayon prelude for parallelization
 mod gpu_job;
 use crate::input::{InputManager, KeyCode};
 use crate::camera::{FreeCamera, CamMode, CameraState, SwitchDelay};
-use crate::bsp::{BspNode, BspStats, Triangle, Frustum, build_bsp, find_node, collect_triangles_in_subtree, create_plane_mesh, create_highlight_mesh, cpu_mesh_to_triangles, traverse_bsp_with_frustum, render_bsp_tree, triangle_center};
+use crate::bsp::{BspNode, BspStats, Triangle, Frustum, build_bsp, find_node, collect_triangles_in_subtree, create_plane_mesh, create_highlight_mesh, cpu_mesh_to_triangles, traverse_bsp_with_frustum, triangle_center};
 
 mod input;
 mod camera;
 mod bsp;
+mod gui;
 // Message types for the channel
 #[derive(Debug)]
 enum Message {
@@ -530,271 +531,25 @@ fn main() -> Result<()> {
 
         // --- GUI ---
         gui.update(&mut frame_input.events.clone(), frame_input.accumulated_time, frame_input.viewport, frame_input.device_pixel_ratio, |ctx| {
-            egui::SidePanel::left("tree").show(ctx, |ui| {
-                ui.heading("BSP Strom");
-                ui.label(format!("Režim: {:?}", mode));
-
-                // Nová sekce pro načtení souboru - přesunuto na začátek
-                ui.separator();
-                ui.heading("Načtení modelu");
-                ui.label("Aktuální model:");
-                ui.label(format!("{}", loaded_file_name));
-
-                // Tlačítko pro načtení souboru
-                if ui.button("📁 Načíst nový model").clicked() {
-                    // Otevření dialogu pro výběr souboru
-                    if let Some(path) = rfd::FileDialog::new()
-                        .add_filter("GLTF/GLB files", &["gltf", "glb"])
-                        .pick_file() {
-                            // Načítání nového souboru na pozadí
-                            file_loading = true;
-                            let path_clone = path.clone();
-                            let file_name_clone = path.file_name().unwrap().to_string_lossy().into_owned();
-                            let tx_gui_clone = tx_gui.clone();
-
-                            // Asynchronní načítání souboru
-                            thread::spawn(move || {
-                                let (new_cpu_mesh, load_status) = load_cpu_mesh(&path_clone);
-
-                                // Odeslání zprávy do hlavního vlákna
-                                let _ = tx_gui_clone.send(Message::NewFile {
-                                    cpu_mesh: new_cpu_mesh,
-                                    file_name: file_name_clone,
-                                    load_status,
-                                    triangles: Vec::new(),
-                                    bsp_tree: BspNode::new_leaf(Vec::new(), 0) // Prázdný uzel jako placeholder
-                                });
-                            });
-                        }
-                }
-
-                // Indikátor načítání
-                if file_loading {
-                    ui.add(
-                        egui::ProgressBar::new(0.0)
-                            .desired_width(ui.available_width())
-                            .text("Načítání modelu a stavba BSP stromu...")
-                            .animate(true)
-                    );
-                }
-
-                // Pokud se strom ještě nestihl zkonstruovat:
-                if bsp_root.is_none() {
-                    ui.separator();
-                    ui.label("Strom se staví…");
-                    // indeterminovaný progress bar
-                    ui.add(
-                        egui::ProgressBar::new(0.0)
-                            .desired_width(ui.available_width())
-                            .animate(true)
-                    );
-                    return; // nic dalšího netaháme, dokud strom neexistuje
-                }
-
-                // tady už víme, že bsp_root.is_some(), takže kreslíme zbytek UI…
-                // Přidáme sekci pro výběr uzlu BSP stromu
-                ui.separator();
-                ui.heading("Struktura BSP stromu");
-                ui.checkbox(&mut show_splitting_plane, "Zobrazit dělící rovinu");
-                
-                // Přidáme nastavení pro vypnutí cullingu
-                ui.separator();
-                ui.heading("Nastavení zobrazení");
-                ui.checkbox(&mut disable_culling, "Zobrazit celý BSP strom (bez cullingu)");
-                if disable_culling {
-                    ui.label("Varování: Zobrazení celého stromu může zpomalit vykreslování.");
-                }
-                
-                // Použijeme scrollovatelné okno pro zobrazení stromu, aby nepřetekl
-                egui::ScrollArea::vertical().show(ui, |ui| {
-                    let root = bsp_root.as_ref().unwrap();
-                    render_bsp_tree(ui, root, &mut selected_node);
-                });
-                
-                // Zobrazíme informace o vybraném uzlu
-                if let Some(node_id) = selected_node {
-                    if let Some(ref root) = bsp_root {
-                        if let Some(node) = find_node(root, node_id) {
-                            ui.separator();
-                            ui.heading("Vybraný uzel");
-                            ui.label(format!("ID: {}", node.id));
-                            ui.label(format!("Trojúhelníků: {}", node.triangles.len()));
-                
-                            if let Some(ref plane) = node.plane {
-                                ui.label("Dělící rovina:");
-                                ui.label(format!("Normála: ({:.2}, {:.2}, {:.2})",
-                                    plane.n.x, plane.n.y, plane.n.z));
-                                ui.label(format!("Vzdálenost: {:.2}", plane.d));
-                            } else {
-                                ui.label("List (bez dělící roviny)");
-                            }
-                
-                            ui.label("Obalový objem:");
-                            ui.label(format!("Min: ({:.2}, {:.2}, {:.2})",
-                                node.bounds.min.x, node.bounds.min.y, node.bounds.min.z));
-                            ui.label(format!("Max: ({:.2}, {:.2}, {:.2})",
-                                node.bounds.max.x, node.bounds.max.y, node.bounds.max.z));
-                
-                            // Přidáno tlačítko pro odznačení
-                            if ui.button("Odznačit").clicked() {
-                                selected_node = None;
-                            }
-                        }
-                    }
-                }
-
-                ui.separator();
-                ui.heading("BSP Statistiky");
-                ui.label(format!("Celkem uzlů: {}", current_stats.total_nodes));
-                ui.label(format!("Celkem trojúhelníků: {}", current_stats.total_triangles));
-                ui.label(format!("Navštíveno uzlů: {}", current_stats.nodes_visited));
-                ui.label(format!("Vykresleno trojúhelníků: {}", current_stats.triangles_rendered));
-                ui.label(format!("Procházka efektivita: {:.1}%",
-                    if current_stats.total_nodes > 0 {
-                        (current_stats.nodes_visited as f32 / current_stats.total_nodes as f32) * 100.0
-                    } else { 0.0 }));
-
-                ui.separator();
-                ui.heading("Mesh Info");
-                ui.label(format!("Vrcholy: {}", cpu_mesh.positions.len()));
-                match &cpu_mesh.indices {
-                    Indices::U32(idx) => ui.label(format!("Indexy (U32): {}", idx.len())),
-                    Indices::U16(idx) => ui.label(format!("Indexy (U16): {}", idx.len())),
-                    _ => ui.label("Indexy: žádné"),
-                };
-
-                ui.separator();
-                ui.heading("Ovládání");
-
-                ui.label("POHYB:");
-                ui.label("• W - Dopředu");
-                ui.label("• S - Dozadu");
-                ui.label("• A - Doleva");
-                ui.label("• D - Doprava");
-                ui.label("• Space - Nahoru");
-                ui.label("• C - Dolů");
-                ui.label(format!("Rychlost: {:.1}", cam.speed));
-
-                ui.separator();
-                ui.label("ROZHLÍŽENÍ:");
-                ui.label("• ↑ - Díváš se nahoru");
-                ui.label("• ↓ - Díváš se dolů");
-                ui.label("• ← - Otočit hlavu doleva");
-                ui.label("• → - Otočit hlavu doprava");
-                ui.label(format!("Rychlost rozhlížení: {:.1}°/s", cam.look_speed * 180.0 / std::f32::consts::PI));
-                ui.add(egui::Slider::new(&mut cam.look_speed, 0.5..=5.0)
-                    .text("Rychlost rozhlížení"));
-
-                ui.separator();
-                ui.label("OSTATNÍ:");
-                ui.label("• F - Přepnout na režim Spectator");
-                ui.label("• G - Přepnout na režim ThirdPerson");
-                ui.label("• Home - Návrat na výchozí pozici");
-                ui.label("• PageUp/PageDown - Upravit rychlost");
-
-                ui.separator();
-                ui.heading("Informace o kameře");
-                ui.label(format!("Aktivní režim: {:?}", mode));
-                
-                // Informace o Spectator kameře
-                ui.collapsing("Spectator kamera", |ui| {
-                    ui.label(format!("Pozice: ({:.1}, {:.1}, {:.1})",
-                        spectator_state.pos.x, spectator_state.pos.y, spectator_state.pos.z));
-                    ui.label(format!("Směr (yaw): {:.1}°", spectator_state.yaw * 180.0 / std::f32::consts::PI));
-                    ui.label(format!("Náklon (pitch): {:.1}°", spectator_state.pitch * 180.0 / std::f32::consts::PI));
-                    ui.label(format!("Rychlost: {:.1}", spectator_state.speed));
-                });
-                
-                // Informace o ThirdPerson kameře
-                ui.collapsing("ThirdPerson kamera", |ui| {
-                    ui.label(format!("Pozice: ({:.1}, {:.1}, {:.1})",
-                        third_person_state.pos.x, third_person_state.pos.y, third_person_state.pos.z));
-                    ui.label(format!("Směr (yaw): {:.1}°", third_person_state.yaw * 180.0 / std::f32::consts::PI));
-                    ui.label(format!("Náklon (pitch): {:.1}°", third_person_state.pitch * 180.0 / std::f32::consts::PI));
-                    ui.label(format!("Rychlost: {:.1}", third_person_state.speed));
-                });
-                
-                // Informace o aktuální kameře
-                ui.label(format!("Aktuální pozice kamery: ({:.1}, {:.1}, {:.1})", 
-                    cam.pos.x, cam.pos.y, cam.pos.z));
-                ui.label(format!("Vzdálenost mezi kamerami: {:.1}", 
-                    (spectator_state.pos - third_person_state.pos).magnitude()));
-                
-                ui.checkbox(&mut show_camera_direction, "Zobrazit směr pohledu kamery");
-
-                // Nová sekce pro načtení souboru
-                ui.separator();
-                ui.heading("Načtení modelu");
-                ui.label("Aktuální model:");
-                ui.label(format!("{}", loaded_file_name));
-
-                // Tlačítko pro načtení souboru
-                if ui.button("Načíst nový model").clicked() {
-                    // Otevření dialogu pro výběr souboru
-                    if let Some(path) = rfd::FileDialog::new()
-                        .add_filter("GLTF/GLB files", &["gltf", "glb"])
-                        .pick_file() {
-                            // Načítání nového souboru na pozadí
-                            file_loading = true;
-                            let path_clone = path.clone();
-                            let file_name_clone = path.file_name().unwrap().to_string_lossy().into_owned();
-                            let tx_gui_clone = tx_gui.clone();
-
-                            // Asynchronní načítání souboru
-                            thread::spawn(move || {
-                                let (new_cpu_mesh, load_status) = load_cpu_mesh(&path_clone);
-
-                                // Odeslání zprávy do hlavního vlákna
-                                let _ = tx_gui_clone.send(Message::NewFile {
-                                    cpu_mesh: new_cpu_mesh,
-                                    file_name: file_name_clone,
-                                    load_status,
-                                    triangles: Vec::new(),
-                                    bsp_tree: BspNode::new_leaf(Vec::new(), 0) // Prázdný uzel jako placeholder
-                                });
-                            });
-                        }
-                }
-
-                // Indikátor načítání
-                if file_loading {
-                    ui.label("Načítání nového modelu...");
-                    // Indeterminovaný progress bar
-                    ui.add(
-                        egui::ProgressBar::new(0.0)
-                            .desired_width(ui.available_width())
-                            .animate(true)
-                    );
-                }
-
-                // Zpráva o úspěšnosti načtení souboru
-                if let Ok(msg) = rx.try_recv() {
-                    match msg {
-                        Message::NewFile { cpu_mesh, file_name, load_status, triangles, bsp_tree } => {
-                            current_cpu_mesh = cpu_mesh;
-                            loaded_file_name = file_name;
-                            file_loading = false;
-
-                            // Převod nového meshe na trojúhelníky
-                            current_triangles = cpu_mesh_to_triangles(&current_cpu_mesh);
-
-                            // Stavba nového BSP stromu
-                            let mut next_id = 0;
-                            bsp_root = Some(build_bsp(&current_triangles, 0, &mut next_id));
-                        },
-                        _ => {}
-                    }
-                }
-
-                // Debug info o načteném souboru
-                ui.label("Debug info:");
-                ui.label(format!("Vrcholy: {}", current_cpu_mesh.positions.len()));
-                match &current_cpu_mesh.indices {
-                    Indices::U32(idx) => ui.label(format!("Indexy (U32): {}", idx.len())),
-                    Indices::U16(idx) => ui.label(format!("Indexy (U16): {}", idx.len())),
-                    _ => ui.label("Indexy: žádné"),
-                };
-            });
+            crate::gui::draw_left_panel(
+                ctx,
+                mode,
+                &mut loaded_file_name,
+                &mut file_loading,
+                &tx_gui,
+                &rx,
+                &mut current_cpu_mesh,
+                &mut current_triangles,
+                &mut bsp_root,
+                &mut selected_node,
+                &mut show_splitting_plane,
+                &mut disable_culling,
+                &mut show_camera_direction,
+                &mut spectator_state,
+                &mut third_person_state,
+                &mut cam,
+                &current_stats,
+            );
         });
 
         // --- ovládání ---
